Reorganized Socket IO handlers
