+++ chat_server/sio/__init__.py
@@ -26,15 +26,12 @@
 # NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE,  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-# Import blueprint here to include it to Web App
-from . import (
-    admin as admin_blueprint,
-    auth as auth_blueprint,
-    chat as chat_blueprint,
-    users as users_blueprint,
-    languages as languages_blueprint,
-    files_api as files_api_blueprint,
-    preferences as preferences_blueprint,
-    personas as personas_blueprint,
-    configs as configs_blueprint,
+# Import used sio server instance and handlers here
+from .server import sio
+from .handlers import (
+    session,
+    stt,
+    tts,
+    translation,
+    user_message as languages_blueprint,
 )
